pyproject.toml
[project]
name = "portfolio-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.30",
    "langchain-google-genai>=2.1.12",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.43",
    "unstructured[md]>=0.18.15",
    "uvicorn>=0.37.0",
]
